+++ mbo/container/limited_map.h
@@ -102,42 +102,42 @@ class LimitedMap final
   // Constructors and assignment from other LimitMap/value types.
 
   template<std::forward_iterator It>
-  requires std::constructible_from<KeyValueType, mbo::types::ForwardIteratorValueType<It>>
+  requires types::ConstructibleFrom<KeyValueType, mbo::types::ForwardIteratorValueType<It>>
   constexpr LimitedMap(It begin, It end, const KeyComp& key_comp = KeyComp()) noexcept
       : LimitedBase(begin, end, key_comp) {}
 
   constexpr LimitedMap(const std::initializer_list<KeyValueType>& list, const KeyComp& key_comp = KeyComp()) noexcept
       : LimitedBase(list, key_comp) {}
 
-  template<std::constructible_from<KeyValueType> U>
+  template<types::ConstructibleInto<KeyValueType> U>
   requires(!std::same_as<U, KeyValueType>)
   constexpr LimitedMap(const std::initializer_list<U>& list, const KeyComp& key_comp = KeyComp()) noexcept
       : LimitedBase(list, key_comp) {}
 
-  template<std::constructible_from<KeyValueType> U, auto OtherN>
+  template<types::ConstructibleInto<KeyValueType> U, auto OtherN>
   requires(MakeLimitedOptions<OtherN>().kCapacity <= kCapacity)
   constexpr LimitedMap& operator=(const std::initializer_list<U>& list) noexcept {
     LimitedBase::operator=(list);
     return *this;
   }
 
-  template<std::constructible_from<Key> OK, std::constructible_from<Value> OV, auto OtherN, typename OtherCompare>
+  template<types::ConstructibleInto<Key> OK, types::ConstructibleInto<Value> OV, auto OtherN, typename OtherCompare>
   requires(MakeLimitedOptions<OtherN>().kCapacity <= kCapacity)
   constexpr explicit LimitedMap(const LimitedMap<OK, OV, OtherN, OtherCompare>& other) noexcept : LimitedBase(other) {}
 
-  template<std::constructible_from<Key> OK, std::constructible_from<Value> OV, auto OtherN, typename OtherCompare>
+  template<types::ConstructibleInto<Key> OK, types::ConstructibleInto<Value> OV, auto OtherN, typename OtherCompare>
   requires(MakeLimitedOptions<OtherN>().kCapacity <= kCapacity)
   constexpr LimitedMap& operator=(const LimitedMap<OK, OV, OtherN, OtherCompare>& other) noexcept {
     LimitedBase::operator=(other);
     return *this;
   }
 
-  template<std::constructible_from<Key> OK, std::constructible_from<Value> OV, auto OtherN, typename OtherCompare>
+  template<types::ConstructibleInto<Key> OK, types::ConstructibleInto<Value> OV, auto OtherN, typename OtherCompare>
   requires(MakeLimitedOptions<OtherN>().kCapacity <= kCapacity)
   constexpr explicit LimitedMap(LimitedMap<OK, OV, OtherN, OtherCompare>&& other) noexcept
       : LimitedBase(std::move(other)) {}
 
-  template<std::constructible_from<Key> OK, std::constructible_from<Value> OV, auto OtherN, typename OtherCompare>
+  template<types::ConstructibleInto<Key> OK, types::ConstructibleInto<Value> OV, auto OtherN, typename OtherCompare>
   requires(MakeLimitedOptions<OtherN>().kCapacity <= kCapacity)
   constexpr LimitedMap& operator=(LimitedMap<OK, OV, OtherN, OtherCompare>&& other) noexcept {
     LimitedBase::operator=(std::move(other));